Rename and add docs
