+++ spec/controllers/concerns/json_api_sort_param_spec.rb
@@ -1,21 +1,21 @@
 require "rails_helper"
 
-RSpec.describe JsonApiSortParams do
-  let!(:controller) { Class.new { include JsonApiSortParams }.new }
+RSpec.describe JsonApiSortParam do
+  let!(:controller) { Class.new { include JsonApiSortParam }.new }
 
   it "returns a default sort order if there are no sorting params" do
-    params = controller.sort_params(
+    params = controller.parse_sort_param(
       "",
-      allowed_params: [:created_at],
+      allowed_fields: [:created_at],
       default_sort: { created_at: :desc },
     )
     expect(params).to eq({ created_at: :desc })
   end
 
   it "filters out non-allowed params", :aggregate_failures do
-    params = controller.sort_params(
+    params = controller.parse_sort_param(
       "-created_at,updated_at",
-      allowed_params: [:updated_at],
+      allowed_fields: [:updated_at],
       default_sort: { created_at: :desc },
     )
     expect(params).not_to have_key(:created_at)
@@ -23,18 +23,18 @@ RSpec.describe JsonApiSortParams do
   end
 
   it "generates a sort params hash" do
-    params = controller.sort_params(
+    params = controller.parse_sort_param(
       "created_at,-updated_at",
-      allowed_params: %i[created_at updated_at],
+      allowed_fields: %i[created_at updated_at],
       default_sort: { created_at: :desc },
     )
     expect(params).to eq({ created_at: :asc, updated_at: :desc })
   end
 
   it "sorts the params in the correct order" do
-    params = controller.sort_params(
+    params = controller.parse_sort_param(
       "-updated_at,created_at",
-      allowed_params: %i[created_at updated_at],
+      allowed_fields: %i[created_at updated_at],
       default_sort: { created_at: :desc },
     )
     expect(params.to_a).to eq([%i[created_at asc], %i[updated_at desc]])
